==> 08_core_microbiome/03_find_shared_taxa.ipynb <==
import pandas as pd
import numpy as np

The idea is to caluclate number and percentage of genera or ASVs that are in 4 groups:  
1) shared across countries, means that there is at least one genus in each country  
2) specific to country - at least one genus in single country, but not in another  
3) specific to region - at least one genus in single region, but now in another  
4) site-specific - at least one genus in a single site only  

# DATA

genus_16s = "02_core_16S_site_GENUS.tsv"
genus_ITS = "05_core_ITS_site_GENUS.tsv"
asv_16s = "02_core_16S_site_ASV.tsv"
asv_ITS = "05_core_ITS_site_ASV.tsv"

pan = ['11']
country_specific = ['10','01']
region_specific = ['100','010','001']
site_specific = ['10000000', '01000000','00100000','00010000','00001000','00000100','00000010','00000001']


# One file example

df = pd.read_csv(genus_ITS, sep="\t", header=0)
df.head()

df = pd.read_csv(asv_ITS, sep="\t", header=0)
print(df.shape)
# Removing genera with no occurences
df['Site'] = df.apply(lambda x: str(x['AMOS'])+str(x['STFE'])+str(x['STET'])+str(x['ESSI'])+str(x['FORE'])+str(x['FLOR1'])+str(x['FP2'])+str(x['Santiago']), axis=1)
df = df[df['Site'] != '00000000'].reset_index(drop=True)
print(df.shape)
# Patterns for countries and regions
df['Canada'] = df.apply(lambda x: max(x['AMOS'],x['STFE'],x['STET'],x['ESSI'],x['FORE']), axis=1)
df['Mexico'] = df.apply(lambda x: max(x['FLOR1'],x['FP2'],x['Santiago']), axis=1)
df['Country'] = df.apply(lambda x: str(x['Canada'])+str(x['Mexico']), axis=1)
df['Boreal'] = df.apply(lambda x: max(x['AMOS'],x['STFE']), axis=1)
df['ColdTemperate'] = df.apply(lambda x: max(x['STET'],x['ESSI'],x['FORE']), axis=1)
df['WarmTemperate'] = df.apply(lambda x: max(x['FLOR1'],x['FP2'],x['Santiago']), axis=1)
df['Region'] = df.apply(lambda x: str(x['Boreal'])+str(x['ColdTemperate'])+str(x['WarmTemperate']), axis=1)
df.head()
df.columns


df['Pan-country_specific_group1'] = df['Country'].apply(lambda x: 1 if x in pan else 0)
df['Country-specific_specific_group2'] = df['Country'].apply(lambda x: 1 if x in country_specific else 0)
df['Canada_spec_group2'] = df['Country'].apply(lambda x: 1 if x in ['10'] else 0)
df['Mexico_spec_group2'] = df['Country'].apply(lambda x: 1 if x in ['01'] else 0)

df['Region-specific_specific_group3'] = df['Region'].apply(lambda x: 1 if x in region_specific else 0)
df['Boreal_spec_group3'] = df['Region'].apply(lambda x: 1 if x in ['100'] else 0)
df['Cold-temperate_spec_group3'] = df['Region'].apply(lambda x: 1 if x in ['010'] else 0)
df['Warm-temperate_spec_group3'] = df['Region'].apply(lambda x: 1 if x in ['001'] else 0)

df['Site-specific_specific_group4'] = df['Site'].apply(lambda x: 1 if x in site_specific else 0)
df['AMOS_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['10000000'] else 0)
df['STFE_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['01000000'] else 0)
df['STET_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00100000'] else 0)
df['ESSI_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00010000'] else 0)
df['FORE_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00001000'] else 0)
df['FLOR1_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00000100'] else 0)
df['FP2_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00000010'] else 0)
df['Santiago_spec_group4'] = df['Site'].apply(lambda x: 1 if x in ['00000001'] else 0)

df.head()

df.columns

potential_columns_to_drop = ['AMOS','ESSI','FLOR1','FORE','FP2','STET','STFE','Santiago','count','genus_abd', 'asv_abd', 'Site', 'Canada', 'Mexico', 'Country',
       'Boreal', 'ColdTemperate', 'WarmTemperate', 'Region','Domain', 'Phylum', 'Class', 'Order', 'Family', 'Genus','Species']
present_columns_to_drop = [ele for ele in potential_columns_to_drop if ele in df.columns]
dsub = df.drop(columns=present_columns_to_drop).reset_index(drop=True)
dstack = dsub.set_index('Row.names').stack(
).reset_index().rename(columns={'level_1':'group',0:'presence'})
dg = dstack.groupby('group').agg(n = ('presence','sum'),
                            total = ('presence', 'count')).reset_index()

dg['percent'] = dg.apply(lambda x: (x['n']/x['total'])*100, axis=1)
dg['label'] = dg['group'].apply(lambda x: x.split('_')[0])
dg['group_id'] = dg['group'].apply(lambda x: x.split('_')[2])
dg['color_group'] = dg['group'].apply(lambda x: x.split('_')[1])

dg

# Loop

genus_16s = "02_core_16S_site_GENUS.tsv"
genus_ITS = "05_core_ITS_site_GENUS.tsv"
asv_16s = "02_core_16S_site_ASV.tsv"
asv_ITS = "05_core_ITS_site_ASV.tsv"

pan = ['11']
country_specific = ['10','01']
pan_region = ['111']
region_specific = ['100','010','001']
pan_site = ['11111111']
site_specific = ['10000000', '01000000','00100000','00010000','00001000','00000100','00000010','00000001']


for filename in [genus_16s, genus_ITS, asv_16s, asv_ITS]:
    print(filename)
    df = pd.read_csv(filename, sep="\t", header=0)
    print(df.shape)
    # Removing genera with no occurences
    df['Site'] = df.apply(lambda x: str(x['AMOS'])+str(x['STFE'])+str(x['STET'])+str(x['ESSI'])+str(x['FORE'])+str(x['FLOR1'])+str(x['FP2'])+str(x['Santiago']), axis=1)
    df = df[df['Site'] != '00000000'].reset_index(drop=True)
    print(df.shape)
    # Patterns for countries and regions
    df['Canada'] = df.apply(lambda x: max(x['AMOS'],x['STFE'],x['STET'],x['ESSI'],x['FORE']), axis=1)
    df['Mexico'] = df.apply(lambda x: max(x['FLOR1'],x['FP2'],x['Santiago']), axis=1)
    df['Country'] = df.apply(lambda x: str(x['Canada'])+str(x['Mexico']), axis=1)
    df['Boreal'] = df.apply(lambda x: max(x['AMOS'],x['STFE']), axis=1)
    df['ColdTemperate'] = df.apply(lambda x: max(x['STET'],x['ESSI'],x['FORE']), axis=1)
    df['WarmTemperate'] = df.apply(lambda x: max(x['FLOR1'],x['FP2'],x['Santiago']), axis=1)
    df['Region'] = df.apply(lambda x: str(x['Boreal'])+str(x['ColdTemperate'])+str(x['WarmTemperate']), axis=1)

    df['Pan-population_shared_Population'] = df['Country'].apply(lambda x: 1 if x in pan else 0)
    df['Country-specific_unique_Population'] = df['Country'].apply(lambda x: 1 if x in country_specific else 0)
    df['Canada_spec_Population'] = df['Country'].apply(lambda x: 1 if x in ['10'] else 0)
    df['Mexico_spec_Population'] = df['Country'].apply(lambda x: 1 if x in ['01'] else 0)

    df['Pan-region_shared_Region'] = df['Region'].apply(lambda x: 1 if x in ['111'] else 0)
    df['Region-specific_unique_Region'] = df['Region'].apply(lambda x: 1 if x in region_specific else 0)
    df['Boreal_spec_Region'] = df['Region'].apply(lambda x: 1 if x in ['100'] else 0)
    df['Cold-temperate_spec_Region'] = df['Region'].apply(lambda x: 1 if x in ['010'] else 0)
    df['Warm-temperate_spec_Region'] = df['Region'].apply(lambda x: 1 if x in ['001'] else 0)

    df['Pan-site_shared_Site'] = df['Site'].apply(lambda x: 1 if x in ['11111111'] else 0)
    df['Site-specific_unique_Site'] = df['Site'].apply(lambda x: 1 if x in site_specific else 0)
    df['AMOS_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['10000000'] else 0)
    df['STFE_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['01000000'] else 0)
    df['STET_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00100000'] else 0)
    df['ESSI_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00010000'] else 0)
    df['FORE_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00001000'] else 0)
    df['FLOR1_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00000100'] else 0)
    df['FP2_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00000010'] else 0)
    df['Santiago_spec_Site'] = df['Site'].apply(lambda x: 1 if x in ['00000001'] else 0)

    potential_columns_to_drop = ['AMOS','ESSI','FLOR1','FORE','FP2','STET','STFE','Santiago','count','genus_abd', 'asv_abd', 'Site', 'Canada', 'Mexico', 'Country',
           'Boreal', 'ColdTemperate', 'WarmTemperate', 'Region','Domain', 'Phylum', 'Class', 'Order', 'Family', 'Genus','Species']
    present_columns_to_drop = [ele for ele in potential_columns_to_drop if ele in df.columns]
    dsub = df.drop(columns=present_columns_to_drop).reset_index(drop=True)
    dstack = dsub.set_index('Row.names').stack(
    ).reset_index().rename(columns={'level_1':'group',0:'presence'})
    dg = dstack.groupby('group').agg(n = ('presence','sum'),
                                total = ('presence', 'count')).reset_index()
    dg['percent'] = dg.apply(lambda x: (x['n']/x['total'])*100, axis=1)
    dg['label'] = dg['group'].apply(lambda x: x.split('_')[0])
    dg['group_id'] = dg['group'].apply(lambda x: x.split('_')[2])
    dg['color_group'] = dg['group'].apply(lambda x: x.split('_')[1])

    # "02_core_16S_site_GENUS.tsv"
    fileID = filename.replace("02_core_","").replace("05_core_","").replace(".tsv","")
    dg.to_csv("11_get_shared_groups_"+fileID+".tsv", sep="\t", header=True, index=False)
